LayoutTests/imported/w3c/web-platform-tests/html/dom/dynamic-markup-insertion/opening-the-input-stream/009-expected.txt


FAIL document.open replacing singleton window assert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton location assert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton history assert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton navigator assert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton applicationCache assert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton sessionStorage assert_equals: expected (undefined) undefined but got (string) "1"
FAIL document.open replacing singleton localStorage assert_equals: expected (undefined) undefined but got (string) "1"
FAIL document.open replacing singleton locationbar assert_equals: expected (undefined) undefined but got (number) 1
